+++ VideoApp/VideoApp/Stores/Auth/Community/CommunityAuthStore.swift
@@ -15,18 +15,18 @@
 //
 
 import Foundation
-import KeychainAccess
 
 class CommunityAuthStore: AuthStoreEverything {
     weak var delegate: AuthStoreWritingDelegate?
-    var isSignedIn: Bool { passcodeStore.passcode != nil }
+    var isSignedIn: Bool { keychainStore.passcode != nil }
     var userDisplayName: String { appSettingsStore.userIdentity }
     private let appSettingsStore: AppSettingsStoreWriting
     private let tokenService: PasscodeTokenService = PasscodeTokenService()
-    private let passcodeStore = PasscodeStore()
+    private let keychainStore: KeychainStoreWriting
 
-    init(appSettingsStore: AppSettingsStoreWriting) {
+    init(appSettingsStore: AppSettingsStoreWriting, keychainStore: KeychainStoreWriting) {
         self.appSettingsStore = appSettingsStore
+        self.keychainStore = keychainStore
     }
 
     func start() {
@@ -43,22 +43,21 @@ class CommunityAuthStore: AuthStoreEverything {
             userIdentity: name,
             roomName: UUID().uuidString
         ) { [weak self] result in
-            // Make sure this is called on main thread
             guard let self = self else { return }
             
             switch result {
             case .success:
                 self.appSettingsStore.userIdentity = name
-                self.passcodeStore.passcode = passcode
+                self.keychainStore.passcode = passcode
                 completion(.success(()))
             case let .failure(error):
                 completion(.failure(error))
             }
         }
     }
-    
+
     func signOut() {
-        passcodeStore.passcode = nil
+        keychainStore.passcode = nil
         appSettingsStore.userIdentity = ""
         delegate?.didSignOut()
     }
@@ -69,103 +68,14 @@ class CommunityAuthStore: AuthStoreEverything {
 
     func fetchTwilioAccessToken(roomName: String, completion: @escaping (String?, Error?) -> Void) {
         tokenService.getToken(
-            passcode: passcodeStore.passcode ?? "", // Change?
+            passcode: keychainStore.passcode ?? "", // Change?
             userIdentity: appSettingsStore.userIdentity,
             roomName: roomName
         ) { result in
-            // Make sure this is called on main thread
-
             switch result {
             case let .success(token): completion(token.token, nil)
             case let .failure(error): completion(nil, error)
             }
         }
     }
-}
-
-class PasscodeTokenService {
-    func getToken(
-        passcode: String,
-        userIdentity: String,
-        roomName: String,
-        completion: @escaping (Result<APITokenResponse, APIError>) -> Void
-    ) {
-        let passcodeComponents = Passcode(fullPasscode: passcode)
-        
-        let session = URLSession.shared
-        let url = URL(string: "https://video-app-\(passcodeComponents.appID)-dev.twil.io/token")!
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        let body = PasscodeRequestInput(passcode: passcodeComponents.passcode, userIdentity: userIdentity, roomName: roomName)
-        let jsonEncoder = JSONEncoder()
-        jsonEncoder.keyEncodingStrategy = .convertToSnakeCase
-        request.httpBody = try! jsonEncoder.encode(body)
-        request.setValue("application/json; charset=utf-8", forHTTPHeaderField: "Content-Type")
-        request.setValue("application/json; charset=utf-8", forHTTPHeaderField: "Accept")
-        
-        let task = session.dataTask(with: request) { data, response, error in
-            // Improve
-            if (response as! HTTPURLResponse).statusCode == 200 {
-                if let data = data {
-                    let jsonDecoder = JSONDecoder()
-                    jsonDecoder.keyDecodingStrategy = .convertFromSnakeCase
-                    let decodedResponse = try! jsonDecoder.decode(APITokenResponse.self, from: data)
-
-                    DispatchQueue.main.async {
-                        completion(.success(decodedResponse))
-                    }
-                }
-            } else {
-                print("Error: \(String(describing: error))")
-            }
-        }
-        
-        task.resume()
-    }
-}
-
-class PasscodeStore {
-    var passcode: String? {
-        get {
-            keychain["passcode"]
-        }
-        set {
-            keychain["passcode"] = newValue
-        }
-    }
-    private let keychain = Keychain()
-}
-
-struct Passcode {
-    let passcode: String
-    let appID: String
-    
-    init(fullPasscode: String) {
-        passcode = String(fullPasscode.prefix(6))
-        appID = String(fullPasscode.dropFirst(6))
-    }
-}
-
-struct PasscodeRequestInput: Codable {
-    let passcode: String
-    let userIdentity: String
-    let roomName: String
-}
-
-enum APIError: Error {
-    case invalidPasscode
-    case other(Error)
-}
-
-struct APITokenResponse: Codable {
-    let token: String
-}
-
-struct APIErrorResponse: Codable {
-    enum ErrorType: String, Codable {
-        case expired
-        case unauthorized
-    }
-    
-    let type: ErrorType
 }
