+++ VideoApp/VideoApp/Stores/Auth/Community/Keychain/KeychainStore.swift
@@ -1,5 +1,5 @@
 //
-//  Copyright (C) 2019 Twilio, Inc.
+//  Copyright (C) 2020 Twilio, Inc.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -16,17 +16,10 @@
 
 import Foundation
 
-class AuthStoreFactory {
-    func makeAuthStore() -> AuthStoreEverything {
-        switch AppInfoStoreFactory().makeAppInfoStore().appInfo.target {
-        case .videoTwilio, .videoInternal:
-            return AhoyAuthStore(
-                api: TwilioVideoAppAPI(),
-                appSettingsStore: AppSettingsStore.shared,
-                firebaseAuthStore: FirebaseAuthStore()
-            )
-        case .videoCommunity:
-            return CommunityAuthStore(appSettingsStore: AppSettingsStore.shared, keychainStore: KeychainStore())
-        }
-    }
+protocol KeychainStoreWriting: AnyObject {
+    var passcode: String? { get set }
+}
+
+class KeychainStore: KeychainStoreWriting {
+    @KeychainStorage(key: "passcode") var passcode: String?
 }
